completed Add and Delete functionality of CRUD

+++ Cumulative1/Models/Teacher.cs
@@ -13,7 +13,7 @@ namespace Cumulative1.Models
         public string TeacherLName { get; set; }
         public DateTime TeacherHireDate { get; set; }
         public string EmployeeNumber { get; set; }
-        public decimal TeacherSalary { get; set; }
+        public float TeacherSalary { get; set; }
        
     }
 }
